remove extraneous qudaEventElapsedTime def, comment out api defs, cast kernel params as int to avoid printf warnings

+++ lib/tune.cpp
@@ -749,8 +749,8 @@ namespace quda
           tunable.checkLaunchParam(param);
           if (verbosity >= QUDA_DEBUG_VERBOSE) {
             printfQuda("About to call tunable.apply block=(%d,%d,%d) grid=(%d,%d,%d) shared_bytes=%d aux=(%d,%d,%d)\n",
-                       param.block.x, param.block.y, param.block.z, param.grid.x, param.grid.y, param.grid.z,
-                       param.shared_bytes, param.aux.x, param.aux.y, param.aux.z);
+                       (int)param.block.x, (int)param.block.y, (int)param.block.z, (int)param.grid.x, (int)param.grid.y, (int)param.grid.z,
+                       (int)param.shared_bytes, (int)param.aux.x, (int)param.aux.y, (int)param.aux.z);
           }
 
           tunable.apply(stream); // do initial call in case we need to jit compile for these parameters or if policy tuning
